refactor: change id to Long

+++ src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -23,7 +23,7 @@ public class ItemController {
      * @return Список объектов {@link ItemDto}, описывающих вещи пользователя.
      */
     @GetMapping()
-    public List<ItemDto> getAllItemsByUserId(@RequestHeader("X-Sharer-User-Id") Integer userId) {
+    public List<ItemDto> getAllItemsByUserId(@RequestHeader("X-Sharer-User-Id") Long userId) {
         return itemService.findAllItemsByUserId(userId);
     }
 
@@ -34,7 +34,7 @@ public class ItemController {
      * @return Объект {@link ItemDto}, описывающий вещь.
      */
     @GetMapping("/{id}")
-    public ItemDto getItem(@PathVariable Integer id) {
+    public ItemDto getItem(@PathVariable Long id) {
         return itemService.findItemById(id);
     }
 
@@ -46,7 +46,7 @@ public class ItemController {
      * @return Список объектов {@link ItemDto}, удовлетворяющих критериям поиска.
      */
     @GetMapping("/search")
-    public List<ItemDto> searchItems(@RequestHeader("X-Sharer-User-Id") Integer userId,
+    public List<ItemDto> searchItems(@RequestHeader("X-Sharer-User-Id") Long userId,
                                      @RequestParam String text) {
         log.info("User: {} search item by string: {}", userId, text);
         return itemService.searchByString(text, userId);
@@ -60,7 +60,7 @@ public class ItemController {
      * @return Объект {@link ItemDto}, представляющий сохраненную вещь.
      */
     @PostMapping
-    public ItemDto saveItem(@RequestHeader("X-Sharer-User-Id") Integer userId,
+    public ItemDto saveItem(@RequestHeader("X-Sharer-User-Id") Long userId,
                             @Valid @RequestBody ItemDto itemDto) {
         log.info("Save item name: {}, owner id: {}", itemDto.getName(), userId);
         return itemService.saveItem(itemDto, userId);
@@ -75,8 +75,8 @@ public class ItemController {
      * @return Объект {@link ItemDto}, представляющий отредактированную вещь.
      */
     @PatchMapping("/{itemId}")
-    public ItemDto patchItem(@RequestHeader("X-Sharer-User-Id") Integer userId,
-                             @PathVariable Integer itemId,
+    public ItemDto patchItem(@RequestHeader("X-Sharer-User-Id") Long userId,
+                             @PathVariable Long itemId,
                              @RequestBody ItemDto itemDto) {
         log.info("Patch item name: {}, owner id: {}", itemDto.getId(), userId);
         return itemService.patchItem(itemDto, itemId, userId);
