+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -10,10 +10,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class Item {
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private Boolean available;
-    private Integer ownerId;
-    private Integer requestId;
+    private Long ownerId;
+    private Long requestId;
 }
